# internal/docs/swagger.json
{
    "components": {
        "schemas": {
            "kvmm_internal_domain.AuthResponse": {
                "properties": {
                    "csrf_token": {
                        "type": "string"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/kvmm_internal_domain.Profile"
                    }
                },
                "type": "object"
            },
            "kvmm_internal_domain.ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "kvmm_internal_domain.FeedResponse": {
                "properties": {
                    "has_more": {
                        "type": "boolean"
                    },
                    "limit": {
                        "type": "integer"
                    },
                    "next_offset": {
                        "type": "integer"
                    },
                    "offset": {
                        "type": "integer"
                    },
                    "posts": {
                        "items": {
                            "$ref": "#/components/schemas/kvmm_internal_domain.Post"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "kvmm_internal_domain.Gender": {
                "enum": [
                    "male",
                    "female",
                    "other"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "GenderMale",
                    "GenderFemale",
                    "GenderOther"
                ]
            },
            "kvmm_internal_domain.LoginRequest": {
                "properties": {
                    "login": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "login",
                    "password"
                ],
                "type": "object"
            },
            "kvmm_internal_domain.MeResponse": {
                "properties": {
                    "csrf_token": {
                        "type": "string"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/kvmm_internal_domain.Profile"
                    }
                },
                "type": "object"
            },
            "kvmm_internal_domain.Post": {
                "properties": {
                    "author": {
                        "$ref": "#/components/schemas/kvmm_internal_domain.Profile"
                    },
                    "author_community_id": {
                        "type": "integer"
                    },
                    "author_profile_id": {
                        "type": "integer"
                    },
                    "comments_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "likes_count": {
                        "type": "integer"
                    },
                    "media_count": {
                        "type": "integer"
                    },
                    "media_urls": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "post_text": {
                        "type": "string"
                    },
                    "reposts_count": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "kvmm_internal_domain.Profile": {
                "properties": {
                    "avatar_urls": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "bio": {
                        "type": "string"
                    },
                    "birthday": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "gender": {
                        "$ref": "#/components/schemas/kvmm_internal_domain.Gender"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "image_url": {
                        "type": "string"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "patronymic": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "profile_name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "kvmm_internal_domain.RegisterRequest": {
                "properties": {
                    "bio": {
                        "maxLength": 256,
                        "type": "string"
                    },
                    "birthday": {
                        "type": "string"
                    },
                    "confirm_password": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "gender": {
                        "$ref": "#/components/schemas/kvmm_internal_domain.Gender"
                    },
                    "nickname": {
                        "maxLength": 32,
                        "minLength": 4,
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 72,
                        "minLength": 8,
                        "type": "string"
                    },
                    "patronymic": {
                        "maxLength": 32,
                        "minLength": 2,
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "profile_name": {
                        "maxLength": 32,
                        "minLength": 2,
                        "type": "string"
                    },
                    "surname": {
                        "maxLength": 32,
                        "minLength": 2,
                        "type": "string"
                    }
                },
                "required": [
                    "confirm_password",
                    "gender",
                    "nickname",
                    "password",
                    "profile_name",
                    "surname"
                ],
                "type": "object"
            }
        },
        "securitySchemes": {
            "SessionCookie": {
                "in": "cookie",
                "name": "session_id",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "API для регистрации, авторизации и последовательной ленты постов",
        "title": "Kvmm API",
        "version": "1.0"
    },
    "externalDocs": {
        "description": "",
        "url": ""
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Авторизует пользователя и создаёт сессию",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/kvmm_internal_domain.LoginRequest",
                                "summary": "request",
                                "description": "Данные авторизации"
                            }
                        }
                    },
                    "description": "Данные авторизации",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.AuthResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "summary": "Авторизация пользователя",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "description": "Завершает текущую сессию пользователя.",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "summary": "Выход из аккаунта",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "description": "Возвращает пользователя по авторизованной сессии",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.MeResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "summary": "Текущий пользователь",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "Регистрирует пользователя и создаёт авторизованную сессию",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/kvmm_internal_domain.RegisterRequest",
                                "summary": "request",
                                "description": "Данные регистрации"
                            }
                        }
                    },
                    "description": "Данные регистрации",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.AuthResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.ErrorResponse"
                                }
                            }
                        },
                        "description": "Conflict"
                    }
                },
                "summary": "Регистрация пользователя",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/posts": {
            "get": {
                "description": "Возвращает посты в обратном хронологическом порядке",
                "parameters": [
                    {
                        "description": "Количество пропускаемых постов",
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Размер страницы (максимум 10)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 10,
                            "maximum": 10,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kvmm_internal_domain.FeedResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Лента публикаций",
                "tags": [
                    "posts"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Локальный backend",
            "url": "http://localhost:8080"
        }
    ]
}